re-working star icons
