+++ packages/ds-medicare-gov/src/components/Icons/RoundedStarIcon.tsx
@@ -1,0 +1,24 @@
+import React, { FunctionComponent } from 'react';
+import { SvgIconProps } from '@cmsgov/design-system/dist/components/Icons/SvgIcon';
+import FilledStar from './RoundedStarFilledIcon';
+import HalfStar from './RoundedStarHalfIcon';
+import EmptyStar from './RoundedStarEmptyIcon';
+
+export type RoundedStarVariation = 'filled' | 'half' | 'empty';
+
+export interface RoundedStarProps extends SvgIconProps {
+  className?: string;
+  variation?: RoundedStarVariation;
+}
+
+const RoundedStar = (props: RoundedStarProps) => {
+  if (props.variation === 'filled') {
+    return <FilledStar className={props.className} {...props} />;
+  } else if (props.variation === 'half') {
+    return <HalfStar className={props.className} {...props} />;
+  } else {
+    return <EmptyStar className={props.className} {...props} />;
+  }
+};
+
+export default RoundedStar;
