+++ packages/ds-medicare-gov/src/components/Icons/index.ts
@@ -2,3 +2,4 @@ export { default as Caret } from './Caret';
 export { default as No } from './No';
 export { default as Close } from './CloseSymbol';
 export { default as Checkmark } from './Checkmark';
+export { default as RoundedStarIcon } from './RoundedStarIcon';
